requirements.txt
numpy==1.21.0
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
pandas==1.3.0
PyYAML==5.4.1
flask==2.0.1
matplotlib==3.4.2